local-ai-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pillow==11.1.0
torch>=2.2.0
torchvision>=0.17.0
accelerate>=0.26.0
safetensors>=0.4.5
qwen-vl-utils==0.0.14
# Install latest Transformers separately with:
# pip install git+https://github.com/huggingface/transformers accelerate
